+++ MoeaBench/core/__init__.py
@@ -5,4 +5,5 @@
 
 from .experiment import experiment as Experiment
 from .run import Run, SmartArray, Population
+from .save import save
 from .base_moea import BaseMoea